Moved autodecorate/help/ to help/
