+++ cyclopts/__init__.py
@@ -64,6 +64,6 @@ from cyclopts.parameter import Parameter
 from cyclopts.protocols import Dispatcher
 from cyclopts.token import Token
 from cyclopts.utils import UNSET, default_name_transform
-from cyclopts.autodecorate.help import CycloptsPanel
+from cyclopts.help import CycloptsPanel
 
 from . import config, types, validators
