+++ tests/test_help.py
@@ -10,7 +10,7 @@ import pytest
 from cyclopts import App, Group, Parameter
 from cyclopts.argument import ArgumentCollection
 from cyclopts.exceptions import CoercionError, MissingArgumentError
-from cyclopts.autodecorate.help import (
+from cyclopts.help import (
     AbstractRichHelpPanel,
     create_parameter_help_panel,
     format_command_entries,
